+++ packages/client/src/components/playtest/SamplePackView.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import { Col, Row } from 'components/base/Layout';
+import { Col, Flexbox, Row } from 'components/base/Layout';
+import CreateP1P1Card from 'components/p1p1/CreateP1P1Card';
 import PreviousP1P1sCard from 'components/p1p1/PreviousP1P1sCard';
 import SamplePackCard from 'components/SamplePackCard';
 
@@ -23,14 +24,17 @@ const SamplePackView: React.FC<SamplePackViewProps> = ({
         <SamplePackCard />
       </Col>
       <Col xs={12} xl={6}>
-        {previousPacks.length > 0 && (
-          <PreviousP1P1sCard
-            packs={previousPacks}
-            packsLastKey={previousPacksLastKey}
-            cubeId={cubeId}
-            cubeOwner={cubeOwnerId}
-          />
-        )}
+        <Flexbox direction="col" gap="2">
+          <CreateP1P1Card cubeId={cubeId} />
+          {previousPacks.length > 0 && (
+            <PreviousP1P1sCard
+              packs={previousPacks}
+              packsLastKey={previousPacksLastKey}
+              cubeId={cubeId}
+              cubeOwner={cubeOwnerId}
+            />
+          )}
+        </Flexbox>
       </Col>
     </Row>
   );